pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "herbalsuite"
version = "0.1.0"
description = "HerbalSuite – Clean-room Seed basierend auf GNUmed-Ideen (ohne Code-Übernahme)."
authors = [{name = "DigitaleAlchemy", email = "info@example.com"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PySide6>=6.6",
  "SQLAlchemy>=2.0",
  "psycopg2-binary>=2.9",
  "pydantic>=2.8",
  "python-dotenv>=1.0",
  "PyYAML>=6.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
herbalsuite = "herbalsuite.app:main"
